python scripts
In this commit, python code was moved to `.py` files in order to
facilitate training from the command line.

Also, two convolutional networks were added. The first implemented
two convolutional layers and two fully connected layers instead of
one. The second extends the first by adding batch normalization to
it (but it is still not fully working -- there is a problem
related to calculating the moving average related to the
normalization when making inference).