predict function
In this commit we added a predict() function to the Model class, in
order to get the final predictions in terms of a score between 0
and 1 that represents a probability.

In addition, we created a notebook (View Predictions.ipynb) to
visualize in a plot how the predictions can be compared with the
ground truth labels for a validation set.